--- backend/ddang/src/main/java/com/ddang/ddang/authentication/application/exception/InvalidTokenException.java
@@ -1,8 +1,0 @@
-package com.ddang.ddang.authentication.application.exception;
-
-public class InvalidTokenException extends IllegalArgumentException {
-
-    public InvalidTokenException(final String message) {
-        super(message);
-    }
-}
