+++ backend/ddang/src/main/java/com/ddang/ddang/authentication/domain/exception/InvalidTokenException.java
@@ -1,0 +1,12 @@
+package com.ddang.ddang.authentication.domain.exception;
+
+public class InvalidTokenException extends IllegalArgumentException {
+
+    public InvalidTokenException(final String message) {
+        super(message);
+    }
+
+    public InvalidTokenException(final String message, final Throwable cause) {
+        super(message, cause);
+    }
+}
