--- backend/ddang/src/main/java/com/ddang/ddang/authentication/infrastructure/oauth2/exception/InvalidSocialOauth2TokenException.java
@@ -1,8 +1,0 @@
-package com.ddang.ddang.authentication.infrastructure.oauth2.exception;
-
-public class InvalidSocialOauth2TokenException extends IllegalArgumentException {
-
-    public InvalidSocialOauth2TokenException(final String message, final Throwable cause) {
-        super(message, cause);
-    }
-}
